requirements.txt
dapr-ext-fastapi==1.7.0
numpy==1.23.1
msgpack-numpy==0.4.8

